+++ tests/pandas_test.py
@@ -26,8 +26,8 @@ def pandas_extension():
     jsonpickle.ext.pandas.unregister_handlers()
 
 
-def roundtrip(obj):
-    return jsonpickle.decode(jsonpickle.encode(obj))
+def roundtrip(obj, **kwargs):
+    return jsonpickle.decode(jsonpickle.encode(obj, **kwargs))
 
 
 def test_series_roundtrip():
@@ -59,6 +59,8 @@ def test_dataframe_roundtrip():
             'an_inf': np.array([np.inf] * 3),
             'a_str': np.str_('foo'),
             'date': np.array([np.datetime64('2014-01-01')] * 3, dtype="datetime64[s]"),
+            'date_ns': np.array([np.datetime64('2014-01-01')] * 3, dtype="datetime64[ns]"),
+            'timedelta': np.array([np.timedelta64(1, "Y")] * 3, dtype="timedelta64[Y]"),
             'complex': np.complex128([1 - 2j, 2 - 1.2j, 3 - 1.3j]),
             # TODO: the following dtypes are not currently supported.
             # 'object': np.object_([{'a': 'b'}]*3),
@@ -67,6 +69,24 @@ def test_dataframe_roundtrip():
     decoded_df = roundtrip(df)
     assert_frame_equal(decoded_df, df)
 
+def test_dataframe_nested_containers():
+    # adapted from #407
+    a = pd.DataFrame({"date": ["20220921abc", 20220921.5], "date2": [20220921, "20220921"]})
+    b = roundtrip(a, keys=True)
+    assert a['date'][0] == b['date'][0]
+
+def test_dataframe_mixed_dtypes():
+    # adapted from #358
+    df = pd.DataFrame({'col1': [[1, 2, 3], ['a', 'b', 'c']]})
+    assert isinstance((df['col1'].iloc[0]), list)
+    result = roundtrip(df)
+    assert isinstance((result['col1'].iloc[0]), list)
+
+    # adapted from #457
+    # Create simple data frame of mixed data types
+    df1 = pd.DataFrame(([1, 2], [4, 'foo']))
+    df2 = roundtrip(df1)
+    assert (df1[0] == df2[0]).all()
 
 def test_multindex_dataframe_roundtrip():
     df = pd.DataFrame(
@@ -301,6 +321,7 @@ def test_multilevel_columns():
     assert isinstance(cloned_data_frame, pd.DataFrame)
     assert data_frame.columns.names == cloned_data_frame.columns.names
     assert_frame_equal(data_frame, cloned_data_frame)
+    assert names == cloned_data_frame.columns.names
 
 
 if __name__ == '__main__':